Revert "Revert "[#FF] refactor: bank account auto-selection hook (#12590)" (#12597)" (#12598)
This reverts commit d9de2d6c22ae1476654aa1e3005f26ce0dd1d8a3.
